Remove test shim, make rolex a module.
Previously the entirety of the program lived in an executable file named
"rolex" (to make it easy to install/run), and rolex.py execfile'd that
to make it importable for tests.

This moves the implementation into "rolex.py", so that there's no test
shim, so installers have to write the file as "rolex" themselves.